refactor: 리팩토링 (리뷰사항 반영)
